Support Node v14 (#30)
* Support Node v14

The minimum (and default) version of Node.js has been changed from v16
to v14. Our libraries still use v14 by default, and `metamask-mobile`
remains on v14 for now (they are in the process of upgrading, but they
require additional time to resolve tooling incompatibilities). The v16
minimum had become a blocker to mobile using this package.

This package and all of its dependencies appear to be compatibile with
Node v14 already.

* Use babel coverage provider; v8 provider is broken in v14